create tools.py
